Use SPDX copyright and license headers
The modern way to specify copyright and license is using SPDX[1]
headers. This is nicer for developers and more friendly for programs
that want to extract this info.

This change was created by the ruse tool[2], using this command:

    reuse init
    ...
    reuse addheader \
        --copyright 'The RamenDR authors' \
        --license Apache-2.0 \
        --exclude-year \
        --skip-unrecognised \
        --template ramen \
        `git ls-files | grep -v config/`

Few unrecognized files were modified manually.

The copyright excludes the years since this info is not really needed,
and this avoids the work of updating all copyrights every year.

The `LICENSE` file was moved to `LICENSES/Apache-2.0.txt`, so we can
check the project with the reuse tool. LICENSE is now a symbolic link to
`LICENSES/Apache-2.0.txt` to make GitHub and Go tools happy.

Some files and directories are special cases and will be handled in a
later commit:
- The config directory, which is partly auto-generated by the operator
  sdk.
- The DCO file, copyrighted by The Linux Foundation but has no license
  info.

[1] https://spdx.dev/
[2] https://git.fsfe.org/reuse/tool

Signed-off-by: Nir Soffer <nsoffer@redhat.com>