Remove the DCO file
The DCO file is not needed to enforce the `Developer Certificate of
Origin` policy.

This file also does not have a valid license info. It only has this note:

    Everyone is permitted to copy and distribute verbatim copies of this
    license document, but changing it is not allowed.

which is not recognized by the reuse tool as valid open source license:

    $ reuse lint
    ...
    The following files have no licensing information:
    * DCO

Instead of wasting time on finding a valid way to license this file,
lets remove it and point to the DOC URL[1].

To enforce DCO policy[2] we can configure the DCO app[3] in the RamenDR
organization. Once configured, any new pull request will include a DCO
workflow enforcing that all commit have a singed-off-by header.

[1] https://developercertificate.org/
[2] https://wiki.linuxfoundation.org/dco#enabling_dco_enforcement
[3] https://github.com/apps/dco
[4] https://github.com/nirs/userstorage/pull/29/checks?check_run_id=9125092601

Signed-off-by: Nir Soffer <nsoffer@redhat.com>