e2e: introduce e2e framework
This commit adds
* e2e directory
* e2e config directory and a config file

We use viper to do the loading and validation of the config.

We have a new go.mod for the e2e dir and therefore we need go.work in
the top dir. This ensures that we keep the testing dependencies separate
from the Ramen dependencies. GO WORK is supported and recommended way of
doing this from Go version 1.18.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>