Add annotations to ProtectedPVC
When we create a protected PVC we need to keep the OCM annotations (and
maybe others) so they exist in the replicationdestination spec. When
restoring a PVC during relocate, we need to include the stored
annotations in the new PVC. This is required to keep the PVC owned by
OCM.

The first step is adding annotation to the type.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>