Extract UpdatePVCLabels helper
The golangci-lint does not allow adding more code in
ensurePVCFromSnapshot() due to complexity. Lets move out the boring
details of updating PVC labels. Hopefully we can use this elsewhere
later.

When pvc.Label is nil, do *not* share the map between the PVC and the
ProtectedPVC. This is very fragile and inconsistent. Now we either
update the existing map or allocating a new map owned by the PVC.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>