Add kubevirt addon and testing environment
The addon installs kubvirt v1.0.0 (corresponds to OpenShift 4.14) in
the managed clusters. We may switch later to latest release, starting
with something more stable.

The environment is minimal environment for testing kubvirt on minikube,
based on kubevirt minikube docs[1].

We use kustomization to make it easy to patch resources with minimal
code changes and easy testing.

[1] https://kubevirt.io/quickstart_minikube/

Signed-off-by: Nir Soffer <nsoffer@redhat.com>