Add secret to test vm
We can generate the secret during kustomization, but it is not possible
to refer to ~/.ssh/ so we copy the user default public key to the vm
kustomization directory.

With this change we can connect to the vm using `virtctl ssh` without
creating the namespace and the secret key manually.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>