Add drenv suspend and resume commands
Suspend or resume the underlying virtual machines. Assuming kvm2 driver
to keep it simple for now, need to implement it better later so it works
also with qemu2 driver.

The use case is building the environment with good network, suspending
it, and resuming it in an environment with flaky network for demo.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>