Improve drenv cache refresh
Previously to update the cache you had to clear the cache and fetch.
This leaves a window where a build may use uncached value and fail.  A
better way is to refresh the cache periodically *without* clearing the
cache, ensuring that builds will always use cached values.

The `fetch` command was renamed to `cache`. This describe better the
purpose of the command. The hook was update as well to match the command
name.

The `drenv.cache` module provides now 2 operations:

- cache.refresh() - rebuild the cache if older than
  cache.REFRESH_THRESHOLD (12 hours). Will be used in the `cache` hook.

- cache.get() - rebuild the cache if older than cache.FETCH_THRESHOLD
  (48 hours) and return the path to the cached value. Will be used in
  `start` hooks to get a cached value.

With this change, we can run `drenv cache envs/regional-dr.yaml` from a
cron job once a day to refresh the cache.

To handle failures, we can retry the command until it succeeds.  Updated
cache items will be refresh only once within the refresh threshold.

The following examples demonstrate the new commands in the lab running
the e2e job.

XXX

Signed-off-by: Nir Soffer <nsoffer@redhat.com>