Add simple gather script
The script gathers all api resources and logs from all or specific
namespaces in a single cluster.

It also runs some ceph commands and copies rbd-mirror log from the node.
This works only for cluster with rook-ceph and when rbd-mirror is
configured to log to file.

Gathering all resources from all namespaces is very slow. It can be much
faster if gather data in parallel.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>