Replace csi-hostpath-driver with rook-cephfs
We added csi-hostpath-driver as a quick temporary solution until we have
cephfs storage. Now that we have it, we can replace it and enjoy reduced
start time, in particular with minikube 1.33.

To replace csi-hostpath-driver, we have to add cephfs to the volsync
development environment. This is slower locally, but faster in the e2e
lab. For regional-dr, this is always faster, up to 1.82 time faster in
the e2e lab.

The main difference is cluster start time - minikube addons are loaded
before minikube start returns.

Before:

    2024-05-12 23:01:42,844 INFO    [dr2] Cluster started in 433.20 seconds
    2024-05-12 23:02:07,215 INFO    [dr1] Cluster started in 457.57 seconds

After:

    2024-05-12 23:18:13,386 INFO    [hub] Cluster started in 71.87 seconds
    2024-05-12 23:18:46,943 INFO    [dr2] Cluster started in 105.43 seconds

Start time before and after this change:

| env          | local before | local after | lab before | lab after |
|--------------|--------------|-------------|------------|-----------|
| regional-dr  |          636 |         426 |        780 |       427 |
| volsync      |          261 |         352 |        520 |       395 |

Signed-off-by: Nir Soffer <nsoffer@redhat.com>