Add external provider
Before this change we had only one provider, so this was internal
implementation detail. This change adds the second provider, allowing
users to configure the provider in the environment file.

Replace the `external: true` option with `provider: external`. With this
we can remove the special handling or external cluster with calls to the
external provider which mostly do nothing.

The external provider may be improved later, for example we may want to
configure containerd for kubevirt. Since we don't have yet environment
using this configuration I'm leaving as is now.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>