Support watching stderr
Some commands (like drenv) log to stderr without writing anything to
stdout. When we watch such commands we want to watch stderr instead of
stdout. Since the command do not write anything to stdout, we can
redirect the command stderr to stdout.

When a command fails, we cannot report the error message since it was
already yielded to the code watching the command. This is the issue with
logging everything to stderr, but we don't control the commands we run.

This change add an option to redirect stderr to commands.watch() and
and test the behavior.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>