Add process_group option to commands.watch()
This is a new feature in python 3.11, allowing running a process in a
new process group. This is required for lima to keep the lima hostagent
process running when drenv shuts down.

This requires python 3.11. We can add the flag later only on macOS so we
can run drenv with older python on Linux, but python 3.11 is available
on Linux for long time.

Lima does not run the hostagent process in the background on Linux or
macOS[1]. When this is fixed we can stop using the process_group option,
but having it is nice anyway.

[1] https://github.com/lima-vm/lima/issues/2573

Signed-off-by: Nir Soffer <nsoffer@redhat.com>