Test lima fix for running in background
With the fix[1] lima starts the hostagenet process in a new process
group so we can run limactl normally as it should. When we terminate
after errors limactl is terminated without harming the hostagenet
process.

The fix is available in upstream but not released yet. It should
available in lima around October.

[1] https://github.com/lima-vm/lima/pull/2574

Signed-off-by: Nir Soffer <nsoffer@redhat.com>