Add setup for macOS
On macOS setup is much simpler since most of the tools are available via
the brew package manager.

For creating vms we use the lima project. We need to build upstream
version to consume important fixes and enhancements. We would be able to
use next lima release around October.

For shared network we use socket_vmnet. It should be installed as root
and the easiest way to do this is building from source.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>