Log drenv log in tests
Like limactl, drenv logs only to stderr, so when running it in tests
with:

    commands.run("drenv", "start", ...)

we don't see anything in the test logs. If the test is blocking for long
time, we have no way to debug this. The helpful log lines are buffered
in the command error buffer.

Use the new stderr= argument to watch and log the command output, and
add helpers for drenv in a consistent way.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>