Modify flannel to use lima0 interface
I used this change when starting the work with lima, and I found that it
is not needed on 2 macs, one connected to a VPN over wired connection, and
the other using WiFi.

Adding it for testing since flannel fails to pull the image from the
registry on Rakesh system, using VPN. We know that the macs connected to
the VPN may have DNS issue.

Thanks: Raghavendra Talur <rtalur@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>