Add natt option
This option is little bit confusing, true by default. We always disable
this on OpenShift, and it is not clear if we need it in the test
environment.

Using natt=False does not change anything, submariner is still broken,
and `subctl show` reports that we *are* using NAT. Keeping the option to
make it easy to modify this in the future.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>