Log limactl errors
Because limactl logs everything to stderr, and we watch stderr to
consume the logs. Since we drop limactl logs when running in normal log
level, when limactl fail we don't have any info on the error, and the
only way to debug limactl error is run with verbose mode.

With this change we extract limactl log level and log errors as drenv
errors.

It would be nice to log also limactl warnings, but limactl use warnings
level for info level message so we cannot log them in normal log level.
These warnings are available when running in verbose mode.

Tested by uninstalling socket_vmnet:

    sudo make uninstall.launchd

With this limaclt will fails to connect to the vmnet socket:

    XXX

Signed-off-by: Nir Soffer <nsoffer@redhat.com>