e2e: Simplify test context management
We had 2 places using a test context - test suites, and action helpers.
To work with both, we store the test context in a global map and had
complicated search mechanism to locate the context in both the parent
and child sub-tests.

Simplify the design by moving the actions wrappers (deploy, enable, ..)
to the test context, and passing a text context to the function running
the test flow.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>