e2e: Introduce test.Context.Validate()
Validate() returns and error if the combination of deployer and workload
is not supported. The error is used as the message for skipping the
test. This cleans up the test and makes it easy to add more tests.

In the future this will also validate if the workload can work on the
clusters. One example is testing virtual machines if clusters does not
have have kubevirt and CDI.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>