e2e: Return more context from kubectl error
Previously if kubectl failed, we logged the command stdout, and return
the command error, which does *not* include stderr output from the
command. This ensure that we don't have any way to debug the failure.

Fix by extracting stderr from the command error, and returning an error
wrapping the command error and including the command stdout and stderr.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>