e2e: Pass test context to utility functions
Instead of passing workload, deployer, log to workloads and deployers
functions, pass the text context providing all of them.

This also remove most accesses to the global util.Ctx.Log for creating a
per-test logger, since we already created the logger in the test
context.

util.DeleteNameapce() is used both in test context and in global
context, so it still accepts a logger.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>