Use uncached client for testing volsync
In the volsync test (and maybe in other tests) we create the k8sClient
using a Manager. This returns a cached client without any sequential
create/get coherence which is impossible to work with a testing context.

The code was full with code doing:

1. Create or update a resource, or call ramen API creating or updating
   a resource
2. Waiting until the resource is eventually visible in the client cache.

We use the same client for creating a volsync.VSHandler, and looks
correct to used a cached client in this case, since using uncached
client can hide bugs in actual ramen code.

Another issue with using the same client for ramen and test code, is
that test code can change the client state, hiding bugs in ramen code.
We don't know about actual problem caused by this sharing yet.

We can fix both issues by creating 2 clients:

1. For volsync.VSHandler, we use k8sManager.GetClient() to get the same
   cachedClient we used before
2. For the testing code, create a new client without a cache, in the
   same way we create client in e2e.

Most of the Eventually calls trying to avoid issues with caching were
replace with a single Get(), keeping the behavior of getting the current
resource from api server, and asserting that the call succeed.

Eventually calls that seems to be only a workaround for the caching
issue were removed.

In few places I was not sure if the Eventually call is not needed and
kept it.

Quick test does not show any issue but running 100 iterations fails in
one of the "VolSync_Handler Ensure PVC from ReplicationDestination When
ReplicationDestination exists with snapshot latestImage" tests. The
failure is similar to #1856 that happens in main.

A nice side effect is speeding up the test. With this change if takes 13
seconds, and in main it takes 20 seconds.

Issues:

- k8sClient should be probably renamed to testClient to make it clear
  that is it special client that should not be used with ramen functions
  accepting a cached client. I kept the current name for now to minimize
  the change.

[1] https://pkg.go.dev/sigs.k8s.io/controller-runtime#hdr-Clients_and_Caches

Signed-off-by: Nir Soffer <nsoffer@redhat.com>