e2e: Simplify workload creation
We treated the deployment as generic deployment and passed many
arguments hard coded in the test suite. This scheme will not work for
adding new workloads. We need to move the workloads details to the
implementation, and provide an easy way to create a workload with the
generic options that can work for any workload.

Add a NewDeployment function accepting a revision (git branch or tag
name), and a PVCSpec. The rest of the argument (deployment name, app
name, path) are constants in the deployment implementation.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>