e2e: Add tests configuration
The tests are defined as a list of yaml objects in the configuration
file. We will generate the Go sub tests from this configuration.

This changes resolve the issue of discovering the tests. The test are
static configuration that is easy to inspect and modify as needed.

Developer can create multiple configuration files to run subset of the
tests, or additional tests that cannot run per-pull-request.

We can use different configurations to do more extensive tests in daily
or weekly runs, and quicker tests that run with every pull request.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>